Cleanup

+++ src/module/echo.c
@@ -4,6 +4,7 @@
 #include "fixtures.h"
 #include "log.h"
 
+#include <stddef.h>
 #include <string.h>
 
 TXVC_DEFAULT_LOG_TAG(echo);
@@ -42,7 +43,7 @@ static int echo_set_tck_period(int tckPeriodNs){
 static bool echo_shift_bits(int numBits, const uint8_t *tmsVector, const uint8_t *tdiVector,
         uint8_t *tdoVector){
     TXVC_UNUSED(tmsVector);
-    memcpy(tdoVector, tdiVector, numBits / 8 + !!(numBits % 8));
+    memcpy(tdoVector, tdiVector, (size_t) (numBits / 8 + !!(numBits % 8)));
     return true;
 }
